+++ InvenTree/order/forms.py
@@ -10,9 +10,16 @@ from django.utils.translation import ugettext_lazy as _
 
 from mptt.fields import TreeNodeChoiceField
 
+from djmoney.forms.fields import MoneyField
+
 from InvenTree.forms import HelperForm
 from InvenTree.fields import RoundingDecimalFormField
 from InvenTree.fields import DatePickerFormField
+
+from InvenTree.helpers import clean_decimal
+
+from common.models import InvenTreeSetting
+from common.forms import MatchItemForm
 
 import part.models
 
@@ -285,3 +292,37 @@ class EditSalesOrderAllocationForm(HelperForm):
             'line',
             'item',
             'quantity']
+
+
+class OrderMatchItemForm(MatchItemForm):
+    """ Override MatchItemForm fields """
+
+    def get_special_field(self, col_guess, row, file_manager):
+        """ Set special fields """
+
+        # set quantity field
+        if 'quantity' in col_guess.lower():
+            return forms.CharField(
+                required=False,
+                widget=forms.NumberInput(attrs={
+                    'name': 'quantity' + str(row['index']),
+                    'class': 'numberinput',
+                    'type': 'number',
+                    'min': '0',
+                    'step': 'any',
+                    'value': clean_decimal(row.get('quantity', '')),
+                })
+            )
+        # set price field
+        elif 'price' in col_guess.lower():
+            return MoneyField(
+                label=_(col_guess),
+                default_currency=InvenTreeSetting.get_setting('INVENTREE_DEFAULT_CURRENCY'),
+                decimal_places=5,
+                max_digits=19,
+                required=False,
+                default_amount=clean_decimal(row.get('purchase_price', '')),
+            )
+
+        # return default
+        return super().get_special_field(col_guess, row, file_manager)
